# [One .NET] <FilterAssemblies/> can skip over @(FrameworkReference)